:sparkles: feat: Show standard error output
When the code is run and standard error output is not empty, the error
output is shown in a separate box below the program output for each
test case. Also, the extension used to mark the code as wrong if
the standard error output wasn't empty even if the output matched
with the expected output. Now, it only checks whether the outputs
match or not.