:sparkles: feat: Keep problems under corresponding contest
When creating a new contest, all the problems are stored in the root
folder. If there are a lot of previous contest files and practice
problems in the root folder already, finding the new contest's files
would be really hard. The root directory also becomes very messy.
This commit solves that problem by scraping the contest name along
with all the other information from the problem page and creating a
folder with that contest name to store the problem in.