:sparkles: feat: View editorial
This commit adds a button besides the submit button to view the problem
editorial. Pressing the button toggles the visibility of the editorial.
When the editorial is requested for the first time for a problem, it is
fetched from CF. If the fetch was successful, then the editorial is
saved in the cache file for that problem so that it is available offline
for subsequent use later on.